refactor(ai): extract colonial-lite naval cases and extend pin size gate
Move planColonialLiteNaval pin bodies into a sibling cases module and
gate the thin contract under the military/naval ≤750-or-cases rule so
Phase 8 residual fat-pin work stays enforced (Refs #3997).
